+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Add `Policy#make_statistics` to allow policies to customize statistics initialization.
+
 ## v0.32.1
 
   - Expose `Async::Container::Controller` `#notify`, `#container_class`, and `#graceful_stop` for testing.